add dynamic timeouts in cli and rule options
We solve the following problem: sometimes users are dealing with large
files, and they would get results if the timeout was larger in such
cases, but they don't want to set a very high value timeout value
globally.
In such cases, we allow the timeout to be a function of the size of the
file and of the timeout passed in the cli (or the default of 5 seconds).

The following options can be added to rules, like in this example below:

```
  options:
    dynamic_timeout: true
    dynamic_timeout_unit_kb: 40
    dynamic_timeout_max_multiplier: 10
```

For this to take effect, the cli parameter
`--allow-rule-timeout-control` must be passed.

In the CLI we can set these values globally using the new flags and
parameters:

- `--dynamic-timeout` flag, to enable dynamic timeouts globally;
- `--dynamic-timeout-unit-kb` to define the size of the blocks used to
  calculate the multiplying factor, in kb (default: 30);
- `--dynamic-timeout-max-multiplier`, to place a bound on the max
  multiplier, in order to avoid very excessive timeouts (default: 20).

When `--allow-rule-timeout-control` is set and a rule has some of these
values set, they take precedence over the global values. Specifically, a
rule can enable (when not globally set) or disable the dynamic timeout
for a target, or override the parameters.