fix: skip '!!_implicit_param!' from idents
These parameters are introduced in encodings (Elixir, OCaml) and they
are not expected to be found in the source files.

However, we were looking for them as if they were normal identifiers,
and the rules containing them were never applicable... hence no
findings.

From now on, it's important to only introduce such special identifiers
that begin with the above prefix, so that they are properly excluded.