fix: generate IL for Rust LetCond
For tainting to work, we treat Rust's LetCond, for example in
code like `if let Some(x) = e {if_branch} else {else_branch}`,
similarly to a LetPattern.

Specifically, the above `If` becomes a `G.Switch` with a condition on
`e`, then a pattern `Some(x) => if_branch` and `G.PatWilcard =>
else_branch`.